back-end/requirements.txt
uvicorn
fastapi
python-dotenv
python-multipart
pyjwt
openai